[modify] 디벨로퍼스 수정사항 반영

+++ app/src/main/java/com/genesis/apps/comm/model/vo/developers/OdometerVO.java
@@ -27,10 +27,10 @@ class OdometerVO extends BaseData {
     private String timestamp;
     @Expose
     @SerializedName("value")
-    private int value;
+    private float value;
     @Expose
     @SerializedName("unit")
-    private int unit;
+    private float unit;
     @Expose
     @SerializedName("date")
     private String date;
